Add files via upload
Phase 5 - Timer/Collision

1. Timer
There should be a delay when the user moves the car from one lane to another lane, before the user can again move the car, i.e. whenever an arrow key is pressed,  it should only move one lane, even if the user keeps pressing it. Think about it, it is very easy to accomplish. 
 
2. Collision
The long awaited collision feature will now be implemented. If the car touches another car, there should be a spark(throw some yellow pixels) and the ending message screen should appear, followed by the ending screen. The game will end. If the arrow key is pressed and there is a car in the next lane, the red car stays in the prev lane. The animation will stop with spark shown followed by the ending message screen. 

3. Score
You have to count all the coins that the user collects and show the current score on the screen somewhere. 

4. Fuel 
After so much ambiguity, finally clarity is here. There will be fuel icons appearing on the road, the car has to collect them to refill its fuel tank to keep moving on the road. The fuel will be maintained by the timer, i.e. you will set a certain time, after which the fuel will reduce, and you show it on the screen in the form of a strip that will change colors from green to yellow to red as the fuel will increase or decrease. Set the limits yourself when the fuel bar will be green or yellow or red. The animation will stop when the fuel ends, followed by the ending message screen followed by the ending screen.

5. Other
Apart from them, you can include loading screens and/or animations between the screens, such as you can open the main screen in the form of an animation, or show the instructions in the form of an animation, or show the ending message screen with animation and call a loading screen between screens, as you wish and feel easy. This is not a requirement. This is just for presentation purposes. 

Here your main game is completed. Till this phase, all your features should be working properly and in coherence. Remember, you have to take care that when the game ends, stack is back to its original state, where it was before the game started, all the interrupts are unhooked properly, so that Afd and other programs run without any error.