Add files via upload
Phase 6 - Music

In this phase, you have to add the music feature in your project.

You will implement this using multitasking. So, the main game should be one process and the background music should be a different process. 

You can produce music by writing on the speaker port. You can either read a music file and write it on the speaker port or just write random numbers on the speaker port, but music should be rhythmic and melodious.

Additional stuff:
To improve the presentation of your project, try adding different sounds for coin collecting, fuel refilling, crash, main animation, quitting and a separate but similar sound for other introduction/instruction/ending screens.